+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This document captures the history of significant changes to the wallet-toolbox 
 The git commit history contains the details but is unable to draw
 attention to changes that materially alter behavior or extend functionality.
 
+## wallet-toolbox 2.0.9
+
+Added support for more currency types.
+
 ## wallet-toolbox 2.0.8
 
 Add check for cross session signAction errors.